+++ src/JirungEngine/MathModule/Vector/Vector2D.cpp
@@ -11,7 +11,7 @@ Vector2D::Vector2D(const Vector2D& vector2d) : x { vector2d.x }, y { vector2d.y 
 
 }
 
-Vector2D::Vector2D(int x, int y) : x { x }, y { y } {
+Vector2D::Vector2D(float x, float y) : x { x }, y { y } {
 
 }
 
@@ -30,14 +30,14 @@ Vector2D Vector2D::operator-(const Vector2D& vector2d) {
     return Vector2D { x - vector2d.x, y - vector2d.y };
 }
 
-int Vector2D::operator*(const Vector2D& vector2d) {
+float Vector2D::operator*(const Vector2D& vector2d) {
     return x*vector2d.x + y*vector2d.y;
 }
 
-Vector2D Vector2D::operator*(int n) {
+Vector2D Vector2D::operator*(float n) {
     return Vector2D { x*n, y*n };
 }
 
-Vector2D Vector2D::operator/(int n) {
+Vector2D Vector2D::operator/(float n) {
     return Vector2D { x/n, y/n };
 }
